split DiffractionDatasetPresenter to fix non-refreshing parameters
