+++ ptychodus/model/data/__init__.py
@@ -1,5 +1,6 @@
 from .core import DataCore, DiffractionDatasetPresenter, ActiveDiffractionPatternPresenter
 from .dataset import ActiveDiffractionDataset
+from .io import DiffractionDatasetInputOutputPresenter
 from .patterns import DiffractionPatternPresenter
 from .settings import DiffractionPatternSettings
 from .sizer import DiffractionPatternSizer
@@ -8,6 +9,7 @@ __all__ = [
     'ActiveDiffractionDataset',
     'ActiveDiffractionPatternPresenter',
     'DataCore',
+    'DiffractionDatasetInputOutputPresenter',
     'DiffractionDatasetPresenter',
     'DiffractionPatternPresenter',
     'DiffractionPatternSettings',
